REFACTOR: health-probe into a separate file
