+++ src/index.js
@@ -1,0 +1,5 @@
+const Health = require('./health-probe');
+
+module.exports = {
+    Health : Health
+}
